fixed age issue

+++ database/migrations/2021_10_12_103834_add_columns_bookings_table.php
@@ -15,8 +15,8 @@ class AddColumnsBookingsTable extends Migration
     {
         Schema::table('bookings', function (Blueprint $table) {
             
-            $table->double('total_price')->nullable()->after('pack_id');
-            $table->double('remaining_price')->nullable()->after('total_price');
+            $table->float('total_price')->nullable()->after('pack_id');
+            $table->float('remaining_price')->nullable()->after('total_price');
             $table->enum('payment_type', [0,1])->comment('0: Full, 1: Half')->default(0);
         });
     }
